change CEKClient into adapter for RequestHandler
