+++ clova-extensions-kit-csharp-web/Startup.cs
@@ -17,7 +17,7 @@ namespace LineDC.CEK
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddTransient<ICEKClient>(sp=> new CEKClient());
+            services.AddTransient<ICEKClient>(sp=> new CEKClient(new SampleRequestHandler()));
             services.AddMvc();
         }
