+++ clova-extensions-kit-csharp/ICEKClient.cs
@@ -1,4 +1,5 @@
 ﻿using System.IO;
+using System.Threading;
 using System.Threading.Tasks;
 using LineDC.CEK.Models;
 
@@ -14,7 +15,17 @@ namespace LineDC.CEK
         /// </summary>
         /// <param name="signatureCEK">Signature Header value</param>
         /// <param name="body">Content Body (Stream)</param>
-        /// <returns>CEK Response</returns>
+        /// <returns>CEK Request</returns>
         Task<CEKRequest> GetRequest(string signatureCEK, Stream body, bool skipValidation = false);
+
+        /// <summary>
+        /// Handle CEK Request with your concrete class extended RequestHandler
+        /// </summary>
+        /// <param name="signatureCEK"></param>
+        /// <param name="body"></param>
+        /// <param name="skipValidation"></param>
+        /// <param name="cancellationToken"></param>
+        /// <returns></returns>
+        Task<CEKResponse> HandleRequestAsync(string signatureCEK, Stream body, bool skipValidation = false, CancellationToken cancellationToken = default(CancellationToken))
     }
 }
